Implemented Skill, amended Package/Report,
Skills are an abstraction around auras; a skill is the main way to use auras - it consists of either a passive, an active or both. The passive is automatically applied to the owner when the cooldown finishes, the active has to be activated from outside.
Package and report now incorporate a skill - the origin, who is sort of the prime mover of a chain of source auras applying new ones.