Implemented package feedback
The send package method now returns the packages "content", a payload, so to speak, that results from unwrapping it, eg. the newly created Aura in the case of Spells.
For this purpose (for the events to be more precise), two new non generic interfaces were necessary; non generic methods were added accordingly.
This feature was required for SpellEffect, which needs its Aura to postpone destruction, until its child is no longer, so the backreferences (Source), would not point to destroyed components.