Implemented trait collection
Keyed collecion of unique, extendable, destructible "traits" - unique provides the embedded key, extendable determines what happens to exisiting values when a duplicate wants to be added, destructible provides the cleanup of the duplicate.
Implemented use of trait collection, but had to neuter effect and aura classes, a revamp is overdue.